# mst: make countPrefixLen match JavaScript's UTF-16-edness