And optimize it a bit, so the Go compiler can avoid some bounds checks.

benchstat:

                     │   before    │                after                │
                     │   sec/op    │   sec/op     vs base                │
    CountPrefixLen-8   78.85n ± 1%   38.45n ± 5%  -51.23% (p=0.000 n=10)

                     │   before   │             after              │
                     │    B/op    │    B/op     vs base            │
    CountPrefixLen-8   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
    ¹ all samples are equal

                     │   before   │             after              │
                     │ allocs/op  │ allocs/op   vs base            │
    CountPrefixLen-8   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
    ¹ all samples are equal