@bnewbold 2023-05-03
We should really be working with bytes everywhere internally in MST. The main motivation for this decision was specifically this prefix length counting, in the context of Unicode characters in MST paths. Javascript does UTF-16, while much of the rest of the world does UTF-8, so you get different counts and different MST structures for the same input.

We belt-and-suspenders'd it by:

* changing the MST node DAG-CBOR schema to be byte array not string, for the key suffixes
* ensuring we do this counting on bytes
* validating the MST paths (aka keys) to be a subset of ASCII

We might relax the constraint on MST paths in the future, so we should try to ensure we could work with hypothetical unicode internally, I think? Doing that informally but not explicitly testing, etc, might be a code smell.

The part that is a hold-out from the refactor is having strings pretty much anywhere in MST internals. I think even the interface the MST exposes to the world should probably be bytes not strings, and let the "repo" package handle the decoding/encoding.