@bradfitz 2023-05-03
Good background, but FWIW this change doesn't change any semantics around unicode. It's purely a performance/cleanup change. 

Go `string` and `[]byte` both just store arbitrary bytes. A Go `string` doesn't mean unicode or utf-8 or anything in particular. The only difference is mutability and whether they're `comparable` (a string supports ==, being a map key, etc)

(https://go.dev/blog/strings says more, if this is new information.)