@bradfitz 2023-05-07
@bnewbold, I've updated this PR to match JavaScript's behavior, while also being faster. Maybe this is more interesting to you now? 😄