@bradfitz 2023-05-07
But warning: this still isn't safe for non-ASCII keys in MST yet. The callers will need to be updated. Notably, this only returns the matching `"k"` value for the prefix length that JavaScript encodes. But this doesn't build such non-ASCII trees correctly yet because you can't just slice a UTF-8 encoded string by this number to get to the right suffix. That'll require updating too. But currently all the MST stuff requires ASCII keys, so I think this is still safe meanwhile.