@bnewbold 2023-05-12
Hey, sorry, I think I miscommunication about this. What we identified previously (back in March) was there there was ambiguity with strings w/r/t UTF-16 and UTF-8 counting. The resolution of this ambiguity was to switch from using "string" to "bytes" in the MST node objects, and to specify that strings should be encoded as UTF-8 at the interface boundary (aka, either at the MST package interface, or even better have the MST package interface take bytes, and the "repo" package take care of UTF-8 encoding/decoding).

So, we don't want to replicate UTF-16 encoding here. Instead, a while back we updated the *Typescript* code to explicitly encode to bytes as UTF-8, and that is what the network/wire/file spec is.

The somewhat confusing `TestPrefixLenWide` is a legacy from when I was reviewing this MST code and discovered the strings problem, and we kept that around to demonstrate the issue.