@bradfitz 2023-05-16
So is the plan for the common prefix length integer to be defined as the common UTF-8 bytes or Unicode codepoints?

Do you worry about prefixes that don't have any JSON representation?

That is, is `commonPrefixLen("fooΓ", "fooΔ")` equal to 4 or 3?

Because if the former, e.g. `"Γ"` and `"Δ"` both encode to two bytes of UTF-8, both starting with the byte `0xCE` but that has no JSON representation as a single byte.

But maybe you're always working with DAG-CBOR or other binary formats and you don't need it to round-trip through a Unicode or UTF-8 string. It'd work in Go but I'm not sure how painful working with binary strings is in JS. Maybe you stay in ArrayBuffer or whatever world enough as it is?