@bnewbold 2023-05-16
Common prefix length is defined as bytes, and strings get encoded as UTF-8, so: UTF-8 bytes.

For now, we effectively limit the repo paths to be a subset of ASCII, so we don't really even need an answer about encoding unicode characters. The string examples you give ("fooΓ", "fooΔ") would both be rejected as invalid repo paths today. But we do want to leave the door open to unicode in repo paths in the future, so we do have an answer.

The answer is that "paths as strings" is a repo-layer concept. When the repo contents are encoded in to an MST tree, the first step is to convert the repo paths in to an MST key, and that key is bytes. So the answer to the above would be "4 bytes common prefix".

The changes in the typescript implementation are described in this PR: https://github.com/bluesky-social/atproto/pull/635

Basically, we switched the MST node object from having "strings" to "byte arrays". I think they use [Uint8Array](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Uint8Array) instead of native strings in that implementation.

A good analogy is something like BoltDB, a key/value store. The keys are bytes, and the function signature looks like:

```
func (b *Bucket) Get(key []byte) []byte
func (b *Bucket) Put(key []byte, value []byte) error
```

It is really common to use strings encoded in UTF-8 as keys, but calling code needs to handle that. Similar, atproto MST trees should have `[]byte` in all the function signatures. The calling code in other packages needs to deal with encoding/decoding of strings to UTF-8 bytes.