@bradfitz 2023-05-16
Gotcha. But repeating what I said before: in Go, a `string` doesn't mean any particular encoding or imply Unicode or imply UTF-8. It's just a read-only view of an immutable `[]byte`. So writing `[]byte` in a signature might be a win for docs, but it's not required for correctness. Sometimes `string` (or a typed variant thereof) is better for other reasons due to it being comparable, immutable, etc.