@bradfitz 2023-05-16
I've reverted this PR back to just the optimization it was previously (with no behavior changes).